Moves 'multibody/constraint' into a private location. Remove old (wrong) references in our Doxygen documentation.
