+++ examples/rod2d/BUILD.bazel
@@ -9,6 +9,19 @@ load(
 package(default_visibility = ["//visibility:private"])
 
 drake_cc_library(
+    name = "constraint_solver",
+    srcs = ["constraint_solver.cc"],
+    hdrs = [
+        "constraint_problem_data.h",
+        "constraint_solver.h",
+    ],
+    deps = [
+        "//common:essential",
+        "//solvers:moby_lcp_solver",
+    ],
+)
+
+drake_cc_library(
     name = "rod2d_state_vector",
     srcs = ["rod2d_state_vector.cc"],
     hdrs = ["rod2d_state_vector.h"],
@@ -45,11 +58,10 @@ drake_cc_library(
     hdrs = [
         "rod2d.h",
     ],
-    visibility = ["//multibody/constraint:__pkg__"],
     deps = [
+        ":constraint_solver",
         ":rod2d_state_vector",
         "//common:essential",
-        "//multibody/constraint",
         "//solvers:mathematical_program",
         "//systems/framework:leaf_system",
     ],
@@ -78,13 +90,23 @@ drake_cc_library(
 # === test/ ===
 
 drake_cc_googletest(
+    name = "constraint_solver_test",
+    deps = [
+        ":constraint_solver",
+        ":rod2d",
+        "//solvers:moby_lcp_solver",
+        "//solvers:unrevised_lemke_solver",
+    ],
+)
+
+drake_cc_googletest(
     name = "rod2d_test",
     deps = [
+        ":constraint_solver",
         ":rod2d",
         "//common:essential",
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:expect_no_throw",
-        "//multibody/constraint",
         "//systems/analysis:simulator",
     ],
 )
