+++ examples/rod2d/constraint_problem_data.h
