+++ examples/rod2d/images/colliding-boxes.png
