+++ examples/rod2d/rod2d.h
@@ -4,9 +4,9 @@
 #include <utility>
 #include <vector>
 
+#include "drake/examples/rod2d/constraint_problem_data.h"
+#include "drake/examples/rod2d/constraint_solver.h"
 #include "drake/examples/rod2d/rod2d_state_vector.h"
-#include "drake/multibody/constraint/constraint_problem_data.h"
-#include "drake/multibody/constraint/constraint_solver.h"
 #include "drake/solvers/moby_lcp_solver.h"
 #include "drake/systems/framework/leaf_system.h"
