+++ examples/rod2d/test/constraint_solver_test.cc
@@ -1,4 +1,4 @@
-#include "drake/multibody/constraint/constraint_solver.h"
+#include "drake/examples/rod2d/constraint_solver.h"
 
 #include <cmath>
 #include <memory>
